fix(community): wake the publish loop on author edits and mod-queue comments (#334 #335)
Same class as #226. storeCommentEdit set _communityUpdateTrigger only on
its purge branch, so content/spoiler/nsfw/deleted edits on approved
comments waited for the publishInterval timer. storeComment never set
it; approved comments wake the loop through the lastPostCid/lastCommentCid
change in calculateLatestUpdateTrigger, but those queries exclude
pendingApproval rows and the mod-queue check only runs after the sleep,
so a comment sent to the mod queue was not published to
modQueue.pageCids.pendingApproval until the next tick.

Set the trigger on every comment edit, and in storeComment when the
comment lands in pending approval.

Moves the #226 test to test/node/community/publications-wake-publish-loop.test.ts
and adds the two cases. Measured at publishInterval 30s: author edit
30,099ms -> 266ms, mod-queue comment 30,101ms -> 147ms.
